+++ src/test.rs
@@ -9,6 +9,7 @@ pub mod fixtures {
     use crate::sections::Section;
     use crate::tasks::priority::Priority;
     use crate::tasks::{DateInfo, Deadline, Duration, Task, Unit};
+    use crate::test_time::FixedTimeProvider;
     use crate::time::{self, FORMAT_DATE};
     use chrono::Duration as ChronoDuration;
 
@@ -21,10 +22,12 @@ pub mod fixtures {
             is_favorite: false,
         }
     }
+
     async fn adjusted_date(days: i64) -> String {
         let config = config().await.with_timezone("America/Vancouver");
-        let tomorrow = time::now(&config).unwrap() + ChronoDuration::days(days);
-        tomorrow.format(FORMAT_DATE).to_string()
+        let base_date = time::today_date(&config).unwrap();
+        let adjusted = base_date + ChronoDuration::days(days);
+        adjusted.format(FORMAT_DATE).to_string()
     }
 
     pub async fn today_task() -> Task {
@@ -112,12 +115,14 @@ pub mod fixtures {
     }
 
     pub async fn config() -> Config {
-        let (tx, mut _rx) = tokio::sync::mpsc::unbounded_channel::<Error>();
+        let (tx, _rx) = tokio::sync::mpsc::unbounded_channel::<Error>();
 
         Config::new("alreadycreated", Some(tx))
             .await
             .expect("Could not generate directory")
             .with_projects(vec![project()])
+            .with_time_provider(time::TimeProviderEnum::Fixed(FixedTimeProvider))
+            .with_timezone("America/Vancouver")
     }
 
     pub fn project() -> Project {
@@ -175,8 +180,9 @@ pub mod fixtures {
 }
 #[cfg(test)]
 pub mod responses {
-    use crate::test::fixtures;
     use crate::{VERSION, time};
+
+    use super::fixtures::config;
 
     pub fn sync() -> String {
         String::from(
@@ -724,7 +730,7 @@ pub mod responses {
     }
 
     async fn today_date() -> String {
-        let config = fixtures::config().await.with_timezone("America/Vancouver");
+        let config = config().await.with_timezone("America/Vancouver");
         time::today_string(&config).unwrap()
     }
 }