+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 - Handle `ImageAttachment` comment/file format
 - Reject invalid config file with error message
 - Ignore `is_deleted: true` attachments/comments
+- Handle blank `token` value in Config
+- Fix testCases to be non timezone-dependent using FixedTimeProvider
 
 ## 2025-05-04 v0.7.4